webhook-handler/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pymongo[srv]==3.12.0
python-dotenv
